# Faster pre-load of ResourceTypes