Faster pre-load of ResourceTypes using reflection to make specific in-bulk methods available to call.

The previous method was **43% slower** than the new one, on the **desktop**. Android should see bigger gains with fewer GC runs.

Time to Load R4 on desktop, one by one (previous method)
```
3.305 s
4.006 s
3.301 s
3.152 s
3.311 s
```
Average 3.41 +/- 0.24seconds

Time to Load R4 on the desktop in bulk by Reflection (New Method)
```
2.345 s
2.293 s
2.410 s
2.382 s
2.438 s
```
Average 2.37 +/- 0.04seconds